security/requirements.txt
# Core dependencies
redis==5.0.1
PyJWT==2.8.0
bleach==6.1.0
python-dateutil==2.8.2
typing-extensions==4.9.0

# Web security
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pyOpenSSL==23.2.0
cryptography==41.0.3

# API and web frameworks
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
starlette==0.27.0

# Testing and validation
pytest==7.4.2
pytest-cov==4.1.0
coverage==7.3.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
colorama==0.4.6 